chore: initial commit
fix(sign-in): removed console logs, and changed input type->password

style(selec-ingredient,select): some style

feat(sing-up): picture field added

feat(sing-up): picture field added

mukodo verzió

+++ src/app/shared/services/cuisine-service.ts
@@ -1,5 +1,5 @@
-import { HttpClient } from '@angular/common/http';
-import { Injectable } from '@angular/core';
+import { HttpClient, HttpParams } from '@angular/common/http';
+import { inject, Injectable } from '@angular/core';
 import { API_URL } from '../../config/config';
 
 import { BehaviorSubject } from 'rxjs';
@@ -15,19 +15,47 @@ export interface ICuisin {
 })
 export class CuisinService {
   cuisines$ = new BehaviorSubject<ICuisin[]>([]);
+  private http: HttpClient = inject(HttpClient);
+  private toastService: ToastService = inject(ToastService);
+  isLoading = false;
+  total = 0;
+  pageSize = 20;
+  constructor() {}
 
-  constructor(
-    private http: HttpClient,
-    private toastService: ToastService,
-  ) {}
+  loadNext(search?: string) {
+    if (this.isLoading) return;
+    if (this.total && this.cuisines$.value.length >= this.total) return;
+    this.isLoading = true;
+    const skip = this.cuisines$.value.length;
+    const take = this.pageSize;
+
+    let params = new HttpParams().set('skip', skip).set('take', take);
+    if (search) params = params.set('search', String(search));
+
+    this.http
+      .get<{ data: ICuisin[]; total: number }>(API_URL + '/cuisines', {
+        params,
+      })
+      .subscribe((resp) => {
+        this.cuisines$.next([...this.cuisines$.value, ...resp.data]);
+        this.total = resp.total;
+        this.isLoading = false;
+        //inf?.done();
+      });
+  }
+
+  reset() {
+    this.cuisines$.next([]);
+    this.total = 0;
+  }
 
   getAll() {
     this.http.get<{ data: ICuisin[]; total: number }>(API_URL + '/cuisines').subscribe({
       next: (resp) => {
         this.cuisines$.next(resp.data);
       },
-      error: (err) => {
-        console.error(err);
+      error: (_err) => {
+        //console.error(err);
         this.toastService.add({ message: 'Szerver hiba', type: 'danger' });
       },
     });
@@ -42,16 +70,17 @@ export class CuisinService {
     });
   }
   delete(id: string) {
-    this.http.delete<{ deleted: boolean }>(API_URL + '/cuisines/' + id).subscribe((resp) => {
-      if (resp.deleted) {
-        const updatedCuisines = this.cuisines$.value.filter((c) => c.id !== id);
+    if (id)
+      this.http.delete<{ deleted: boolean }>(API_URL + '/cuisines/' + id).subscribe((resp) => {
+        if (resp.deleted) {
+          const updatedCuisines = this.cuisines$.value.filter((c) => c.id !== id);
 
-        this.cuisines$.next(updatedCuisines);
-        this.toastService.add({ message: 'Sikeresen törölted a kategóriát!', type: 'primary' });
-      } else {
-        this.toastService.add({ message: 'Hiba történt', type: 'danger' });
-      }
-    });
+          this.cuisines$.next(updatedCuisines);
+          this.toastService.add({ message: 'Sikeresen törölted a kategóriát!', type: 'primary' });
+        } else {
+          this.toastService.add({ message: 'Hiba történt', type: 'danger' });
+        }
+      });
   }
   update(cuisin: ICuisin) {
     this.http
